Moved remaining classes of interest from the deprecated client-common module
