--- client-common/src/main/java/eu/arrowhead/client/common/exception/ExceptionType.java
@@ -1,14 +1,0 @@
-/*
- *  Copyright (c) 2018 AITIA International Inc.
- *
- *  This work is part of the Productive 4.0 innovation project, which receives grants from the
- *  European Commissions H2020 research and innovation programme, ECSEL Joint Undertaking
- *  (project no. 737459), the free state of Saxony, the German Federal Ministry of Education and
- *  national funding authorities from involved countries.
- */
-
-package eu.arrowhead.client.common.exception;
-
-public enum ExceptionType {
-  ARROWHEAD, AUTH, BAD_METHOD, BAD_PAYLOAD, BAD_URI, DATA_NOT_FOUND, DNSSD, DUPLICATE_ENTRY, GENERIC, JSON_PROCESSING, UNAVAILABLE, VALIDATION
-}
