--- client-common/src/main/java/eu/arrowhead/client/common/model/OrchestratorWarnings.java
@@ -1,14 +1,0 @@
-/*
- *  Copyright (c) 2018 AITIA International Inc.
- *
- *  This work is part of the Productive 4.0 innovation project, which receives grants from the
- *  European Commissions H2020 research and innovation programme, ECSEL Joint Undertaking
- *  (project no. 737459), the free state of Saxony, the German Federal Ministry of Education and
- *  national funding authorities from involved countries.
- */
-
-package eu.arrowhead.client.common.model;
-
-public enum OrchestratorWarnings {
-  FROM_OTHER_CLOUD, TTL_EXPIRING, TTL_EXPIRED, TTL_UNKNOWN
-}
