+++ common/src/main/java/eu/arrowhead/common/filter/EmptyPayloadFilter.java
@@ -7,10 +7,11 @@
  *  national funding authorities from involved countries.
  */
 
-package eu.arrowhead.client.common.filter;
+package eu.arrowhead.common.filter;
 
-import eu.arrowhead.client.common.exception.ErrorMessage;
-import eu.arrowhead.client.common.exception.ExceptionType;
+import eu.arrowhead.common.exception.ErrorMessage;
+import eu.arrowhead.common.exception.ExceptionType;
+
 import javax.annotation.Priority;
 import javax.ws.rs.Priorities;
 import javax.ws.rs.container.ContainerRequestContext;
