+++ common/src/main/java/eu/arrowhead/common/filter/OutboundDebugFilter.java
@@ -7,9 +7,10 @@
  *  national funding authorities from involved countries.
  */
 
-package eu.arrowhead.client.common.filter;
+package eu.arrowhead.common.filter;
 
-import eu.arrowhead.client.common.Utility;
+import eu.arrowhead.common.misc.Utility;
+
 import javax.annotation.Priority;
 import javax.ws.rs.Priorities;
 import javax.ws.rs.container.ContainerRequestContext;
